[feat/PR7-997] 적립금 적립 및 회원 등급 갱신 로직 구현

+++ user-service/src/main/java/com/princesses7/findy/user/user/dto/response/MyPageResponse.java
@@ -8,7 +8,7 @@ public record MyPageResponse(
 	String name,
 	String email,
 	Grade grade,
-	double reward
+	long reward
 ) {
 
 	public static MyPageResponse from(UserEntity user) {
